test: Split models.recognition unittests
